pyproject.toml
[project]
name = "spark-labs"
version = "1.0.0"
description = "A collection of hands-on labs for learning Apache Spark and big data analytics on cloud computing infrastructures."
readme = "README.md"
requires-python = ">=3.10, <3.11"
dependencies = [
  "jupyter>=1.1.1",
  "pyspark-client==4.0.1",
  "sparkdl>=0.2.2",
]

[tool.ruff]
indent-width = 2